refactor change repositories files for kebak-case
